+++ src/routes/downloads.js
@@ -16,10 +16,10 @@ var vars = require("../variables");
 
 const DownloadableItemCard = ({ downloadableItem, onEdit, onDelete, onDownload }) => {
     const { t: tr } = useTranslation();
-    const { curUserPerm } = useContext(AppContext);
+    const { curUID, curUserPerm } = useContext(AppContext);
 
     const showButtons = onEdit !== undefined;
-    const showControls = (onEdit !== undefined) && (vars.isLoggedIn && checkUserPerm(curUserPerm, ["administrator", "manage_downloads"]));
+    const showControls = (onEdit !== undefined) && (curUID !== null && checkUserPerm(curUserPerm, ["administrator", "manage_downloads"]));
 
     const [isShiftPressed, setIsShiftPressed] = useState(false);
     const [downloading, setDownloading] = useState(false);
@@ -262,7 +262,7 @@ const DownloadableItemManagers = memo(() => {
 
 const DownloadableItem = () => {
     const { t: tr } = useTranslation();
-    const { curUser, curUserPerm } = useContext(AppContext);
+    const { curUID, curUser, curUserPerm } = useContext(AppContext);
 
     const [downloadableItems, setDownloadableItems] = useState([]);
     const [lastUpdate, setLastUpdate] = useState(0);
@@ -307,7 +307,7 @@ const DownloadableItem = () => {
         let url = `${vars.dhpath}/downloads/list?page_size=10&page=${page}`;
 
         var newDowns = [];
-        if (vars.isLoggedIn) {
+        if (curUID !== null) {
             const [anns] = await makeRequestsWithAuth([
                 url
             ]);